vizier_backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
dj-database-url==2.1.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# AWS & Storage
boto3>=1.26.0

# Authentication & JWT
PyJWT>=2.8.0
cryptography>=41.0.0

# Medical Imaging
pydicom>=2.4.0
nibabel>=5.2.0
opencv-python-headless
numpy>=1.24.0
scipy>=1.11.0

# HTTP & API
requests>=2.31.0
httpx>=0.25.0

# Database
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
factory-boy>=3.3.0

# Code Quality & Linting
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
mypy>=1.7.0

# Logging & Monitoring
python-json-logger>=2.0.0

# Utilities
celery>=5.3.0
redis>=5.0.0
